RT #18472: "parl foo.par script.pl ..." uses a cache area that doesn't depend on foo.par
Indeed, the (name of) the cache area used for "parl foo.par ..." is derived from
the SHA1 checksum of parl.exe, and _not_ foo.par. This patch changes that
for the common use case above. It is not a complete fix though, as this would
require to parse the parl command line arguments on the C level.