RT #63083: pp: Undefined subroutine &DynaLoader::bootstrap called in pp-compiled bianry see also RT #38217: script looking for libperl.so.5.8, but libperl.so.5.8.8 exists
myldr: When $par_exe is linked against $libperl the linker writes
a DT_NEEDED tag into $par_exe. But the value of this tag may be not
the same as basename($libperl). The latter is used as the name the packed
copy of $libperl is extracted to. Example from the ticket:

- $libperl was /usr/lib64/libperl.so.5.8.8
- DT_NEEDED was libperl.so.5.8

Hence $libperl is extracted as libperl.so.5.8.8, but the dynamic loader
looks for libperl.so.5.8. Since it isn't found in the cache area,
the loader continues to search the built-in library search path,
finds /usr/lib64/libperl.so.5.8. On another machine
(not the one where PAR::Packer was built) this is actually
a symlink to /usr/lib64/libperl.so.5.8.9. Apparently Perl 5.8.8
and 5.8.9 are slightly ABI incompatible, hence the strange error message.

Solution: Try to use the "correct" name when extracting $libperl.
At least on Linux we should be able to glean this as $libperl's SONAME
using objdump from GNU binutils.

Add this heuristic to myldr/Makefile.PL. Add an option to myldr/file2c.pl
to set the extraction name (and clean it up a bit while we're at it).