Fix RT#104635 for PAR-Packer: IO error the first time an executable built by the packer is run on OSX
The problem isn't specific to OSX - a case-insensitive filesystem is required
and a rare combination of the file names of modules and scripts.
If you run the example given in the bug report, the resulting broken.exe 
contains the following members when viewed as a zip file:

  lib/foo  (a copy of the script that was packed)
  lib/Foo/Bar.pm

When the latter is extracted, a directory lib/Foo will be implicitly
created - and this can't coexist with the file lib/foo on a case-insensitive
filesystem. Note that member lib/foo only shows up if you use
  
  pp eg/foo ...

but not if you move eg/foo to the cwd and do

  pp foo ...

Indeed, comparing the output of 

  scandeps.pl -V eg/foo
  scandeps.pl -V foo

we see that the former list foo as an additional dependency of type "data".
Hence, one can argue that the bug is actually in Module::ScanDeps.
Indeed, at the end of Module::ScanDeps::scan_deps() we find these lines:

  # do not include the input files themselves as dependencies!
  $args{rv}{$_} foreach @{$args{files}};

Note that this only works if an element of $args{files} (of type "data",
esp. a script) does not contain any slashes: its key in $args{rv} 
is the basename of the element, not the filename as given in $args{files}.
In the example, "eg/foo" and "foo" (in $args{files}) will both map 
to key "foo" in $args{rv}. If I make the comment actually work 
(by matching on the full pathname in $args{rv}->{file}), this change
causes lots of tests for Module::ScanDeps to fail. Obviously the current
behaviour is the blessed one, the comment not withstanding.

So I fixed this in PAR::Packer instead: delete copies of the actual
input files under "lib" in the manifest structures.