drop the "re-exec" myself in myldr/internals.c
I dunno what problem this was meant to solve (code has been there since rev 1).
But this execvp may be the cause of test failures on machines with
low argv+env limits: a copy of script/par.pl has been shoved into fakeargv
at the time this is called. It resides in process memory (courtesy of
my_par_pl.c), but execvp pushes it into argv+env territory. It's about ~20 KB,
but that may be too much for some configurations.
Anyway, the execvp() is only relevant for standalone usage of parl
as it's only called when PAR_TEMP hasn't been set up already.
When invoking a packed executable, PAR_TEMP is set up by the boot 
program, before exec'ing parl.