Fix argument list passed to spawnvpe() call on Windows
The first argument in the list should be the program name (or path) being
spawned (i.e. the same as the second argument passed to spawnvpe() itself:
my_perl, in this case); the remainder of the argument list are the
arguments for the spawned process. See:
https://msdn.microsoft.com/en-us/library/20y988d2.aspx#Anchor_1
https://msdn.microsoft.com/en-us/library/h565xwht.aspx#Anchor_0

Also, we can simply use spawnvp() rather than spawnvpe() since the
environment being passed to the spawned process is only the current
process's environ, which the spawned process inherits anyway in the case
of spawnvp(). See:
https://msdn.microsoft.com/en-us/library/20y988d2.aspx#Anchor_2

This also avoids referencing environ, which stdlib.h #defines as _environ,
which is not exposed if _CRT_BEST_PRACTICES_USAGE is #defined (in the
Windows 10 SDK used by VS2015).