Fix #66: patch myldr/boot for "pp --clean ..." without side effects
- make sure myldr/boot contains exactly one string of the form
  "__PASS_PAR_CLEAN__               \0" so that there are no
  duplicates that may get split on chunk boundaries
  (myldr/boot_embedded_files.c)
- simplify patching of this string (in $loader) to
  "__PASS_PAR_CLEAN__=1             \0" in script/par.pl

The probably cause for the original #62 is the following:
A packed executable actually contained 3 occurences of the string
"__ENV_PAR_CLEAN__              \n":
(1) in myldr/boot (#include'd from myldr/mktmpdir.c in myldr/boot.c)
(2) in myldr/par (#include'd from myldr/mktmpdir.c in myldr/main.c)
(3) in myldr/boot via a copy of script/par.pl (as a single C char array)

(2) is embedded in myldr/boot in several chunks as C char arrays (which
concatenated give myldr/par). If the string straddles a chunk boundary
the trailing blanks (in the front part of the string) may not be sufficient
to store the replacement ("PAR_CLEAN=1") The replacement code
in script/par.pl didn't check for this condition, resulting in a
replacement **longer** than the match.  Depending on the
object file layout of myldr/boot this may render the executable invalid.