par_mktmpdir(): check for PAR magic bytes first and fix par.pl
The check that the executable should end with the PAR magic bytes
("\nPAR.pm\n") looks innocent enough, but causes the executables
myldr/par and myldr/boot to contain the string "\nPAR.pm\n" somewhere.
This throws off find_par_magic() when applied to an executable that is
NOT packed by pp, e.g. in PAR::StrippedPARL::Static::write_parl()
where it is effectively applied to a copy of myldr/boot. The resulting
parl then contains a truncated myldr/boot that segfaults when invoked.
You may check with

$ perl -Mblib -MPAR::StrippedPARL::Static \
  -E 'PAR::StrippedPARL::Static->write_parl("/tmp/boot");
      truncate("/tmp/boot", -s "myldr/boot");'
$ perl -Mblib -MPAR::StrippedPARL::Dynamic \
  -E 'PAR::StrippedPARL::Dynamic->write_parl("/tmp/par");
      truncate("/tmp/par", -s "myldr/par");'

The files /tmp/boot and /tmp/par should be identical to myldr/boot and
myldr/par, resp.

The fix to par.pl is to check that the executable also contains "\0CACHE"
at an offset of 4 bytes below the PAR magic. Otherwise pretend we didn't
find the PAR magic.