set up GitHub CI
All straight forward except for this little glitch when
building for perl 5.36 and 5.38:

  g++ main.o sha1.o ppresource.coff ... -o par.exe

fails with
  c:/programdata/chocolatey/lib/mingw/tools/install/mingw64/x86_64-w64-mingw32/bin/ld.exe:
    c:/programdata/chocolatey/lib/mingw/tools/install/mingw64/x86_64-w64-mingw32/lib/crt2.o:crtexe.c:(.text+0x5e):
    undefined reference to `__p__commode'

The reason seems to be that ld.exe from c:/programdata/chocolatey/bin
uses a version of the mingw64 tool chain that is incompatible with the
one thta comes with Strawberry (for 5.36 and 5.38).
c:/programdata/chocolatey/bin comes *before* c:\strawberry\c\bin in
PATH. Hence the hack to add the Strawberry stuff to the fron of PATH
(though shogo82148/actions-setup-perl should have done that already).